test: move dhis-service-administration tests into dhis-test-integration
